Make &AMP; in docs file work
The documents sometimes contain a `&` symbol escaped as `&AMP;`. However, this is then parsed by the XML parser as a marker, which is not defined and therefore the XML parser crashes.

The fix loads the text file, replaces the `&AMP;` by an ` and ` text and then parses the entire XML file.